Fixed navigation bar appearance issues on iOS 26 when navigating between screens, and refactored ShareItemViewModel into smaller, focused view models
Affected files:
- CustomNavigationView.swift
- BaseViewController.swift
- RootNavigationController.swift
- NavigationController.swift
- SettingsScreenView.swift
- ShareItemViewModel.swift
- ShareItemArchivesViewModel.swift (new)
- ShareItemGrantAndInviteViewModel.swift (new)
- ShareItemLinkLifecycleViewModel.swift (new)
- ShareItemLinkSettingsViewModel.swift (new)
