+++ Permanent/Modules/SideMenus/Views/SettingsScreenView.swift
@@ -195,3 +195,5 @@ struct SettingsScreenView: View {
         presentationMode.wrappedValue.dismiss()
     }
 }
+
+
